[FIX] account: deny reset to draft when deferred revenue/expense is opened
Create a deferred revenue/expense
Create an invocie which map to the deferred revenue/expense account
Post the invoice so it will generate the deferred entry (in draft)
Confirm the deferred Entry
Reset to draft the invoice

User will be unable to post the invoice again because of the deferred
revenue/expense open but also unable to access it because the invoice is
not in posted state (deadlock).

Adding a condition to block 'reset to draft' button if there is some
deferred revenue/expense not in draft.

opw-2180330

+++ addons/account/models/account_move.py
@@ -2165,6 +2165,8 @@ class AccountMove(models.Model):
                 raise UserError(_('You cannot reset to draft a tax cash basis journal entry.'))
             if move.restrict_mode_hash_table and move.state == 'posted' and move.id not in excluded_move_ids:
                 raise UserError(_('You cannot modify a posted entry of this journal because it is in strict mode.'))
+            if any(asset_id.state != 'draft' for asset_id in move.asset_ids):
+                raise UserError(_('You cannot reset to draft an entry having a posted deferred revenue/expense'))
             # We remove all the analytics entries for this journal
             move.mapped('line_ids.analytic_line_ids').unlink()
